@@ -5,7 +5,6 @@ import android.net.Uri;
 import com.zhy.http.okhttp.request.GetRequest;
 import com.zhy.http.okhttp.request.RequestCall;
 
-import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Set;
@@ -13,31 +12,23 @@ import java.util.Set;
 /**
  * Created by zhy on 15/12/14.
  */
-public class GetBuilder extends OkHttpRequestBuilder<GetBuilder> implements HasParamsable
-{
+public class GetBuilder extends OkHttpRequestBuilder<GetBuilder> implements HasParamsable {
     @Override
-    public RequestCall build()
-    {
-        if (params != null)
-        {
+    public RequestCall build() {
+        if (params != null) {
             url = appendParams(url, params);
         }
 
-        return new GetRequest(url, tag, params, headers,id).build();
+        return new GetRequest(url, tag, params, headers, id).build();
     }
 
-    protected String appendParams(String url, Map<String, String> params)
-    {
-        if (url == null || params == null || params.isEmpty())
-        {
+    protected String appendParams(String url, Map<String, String> params) {
+        if (url == null || params == null || params.isEmpty()) {
             return url;
         }
         Uri.Builder builder = Uri.parse(url).buildUpon();
         Set<String> keys = params.keySet();
-        Iterator<String> iterator = keys.iterator();
-        while (iterator.hasNext())
-        {
-            String key = iterator.next();
+        for (String key : keys) {
             builder.appendQueryParameter(key, params.get(key));
         }
         return builder.build().toString();
@@ -45,17 +36,14 @@ public class GetBuilder extends OkHttpRequestBuilder<GetBuilder> implements HasP
 
 
     @Override
-    public GetBuilder params(Map<String, String> params)
-    {
+    public GetBuilder params(Map<String, String> params) {
         this.params = params;
         return this;
     }
 
     @Override
-    public GetBuilder addParams(String key, String val)
-    {
-        if (this.params == null)
-        {
+    public GetBuilder addParams(String key, String val) {
+        if (this.params == null) {
             params = new LinkedHashMap<>();
         }
         params.put(key, val);
