@@ -3,13 +3,12 @@ package com.zhy.http.okhttp.callback;
 import okhttp3.Call;
 import okhttp3.Request;
 import okhttp3.Response;
-
+@SuppressWarnings("ALL")
 public abstract class Callback<T>
 {
     /**
      * UI Thread
      *
-     * @param request
      */
     public void onBefore(Request request, int id)
     {
@@ -18,7 +17,6 @@ public abstract class Callback<T>
     /**
      * UI Thread
      *
-     * @param
      */
     public void onAfter(int id)
     {
@@ -27,7 +25,6 @@ public abstract class Callback<T>
     /**
      * UI Thread
      *
-     * @param progress
      */
     public void inProgress(float progress, long total , int id)
     {
@@ -37,8 +34,6 @@ public abstract class Callback<T>
     /**
      * if you parse reponse code in parseNetworkResponse, you should make this method return true.
      *
-     * @param response
-     * @return
      */
     public boolean validateReponse(Response response, int id)
     {
@@ -48,7 +43,6 @@ public abstract class Callback<T>
     /**
      * Thread Pool Thread
      *
-     * @param response
      */
     public abstract T parseNetworkResponse(Response response, int id) throws Exception;
 
