@@ -8,7 +8,7 @@ import okhttp3.Response;
 /**
  * Created by JimGong on 2016/6/23.
  */
-
+@SuppressWarnings("ALL")
 public abstract class GenericsCallback<T> extends Callback<T> {
     IGenericsSerializator mGenericsSerializator;
 
@@ -16,6 +16,7 @@ public abstract class GenericsCallback<T> extends Callback<T> {
         mGenericsSerializator = serializator;
     }
 
+    @SuppressWarnings({"unchecked", "ConstantConditions"})
     @Override
     public T parseNetworkResponse(Response response, int id) throws IOException {
         String string = response.body().string();
@@ -23,8 +24,7 @@ public abstract class GenericsCallback<T> extends Callback<T> {
         if (entityClass == String.class) {
             return (T) string;
         }
-        T bean = mGenericsSerializator.transform(string, entityClass);
-        return bean;
+        return mGenericsSerializator.transform(string, entityClass);
     }
 
 }
