@@ -8,11 +8,11 @@ import java.io.Serializable;
 import okhttp3.Cookie;
 
 /**
- * from http://stackoverflow.com/questions/25461792/persistent-cookie-store-using-okhttp-2-on-android
+ * from <a href="http://stackoverflow.com/questions/25461792/persistent-cookie-store-using-okhttp-2-on-android">...</a>
  * and<br/>
- * http://www.geebr.com/post/okHttp3%E4%B9%8BCookies%E7%AE%A1%E7%90%86%E5%8F%8A%E6%8C%81%E4%B9%85%E5%8C%96
+ * <a href="http://www.geebr.com/post/okHttp3%E4%B9%8BCookies%E7%AE%A1%E7%90%86%E5%8F%8A%E6%8C%81%E4%B9%85%E5%8C%96">...</a>
  */
-
+@SuppressWarnings("ALL")
 public class SerializableHttpCookie implements Serializable
 {
     private static final long serialVersionUID = 6374381323722046732L;
