@@ -11,40 +11,35 @@ import okhttp3.RequestBody;
 /**
  * Created by zhy on 15/12/14.
  */
-public class PostStringRequest extends OkHttpRequest
-{
+@SuppressWarnings("ALL")
+public class PostStringRequest extends OkHttpRequest {
     private static MediaType MEDIA_TYPE_PLAIN = MediaType.parse("text/plain;charset=utf-8");
 
     private String content;
     private MediaType mediaType;
 
 
-    public PostStringRequest(String url, Object tag, Map<String, String> params, Map<String, String> headers, String content, MediaType mediaType,int id)
-    {
-        super(url, tag, params, headers,id);
+    public PostStringRequest(String url, Object tag, Map<String, String> params, Map<String, String> headers, String content, MediaType mediaType, int id) {
+        super(url, tag, params, headers, id);
         this.content = content;
         this.mediaType = mediaType;
 
-        if (this.content == null)
-        {
+        if (this.content == null) {
             Exceptions.illegalArgument("the content can not be null !");
         }
-        if (this.mediaType == null)
-        {
+        if (this.mediaType == null) {
             this.mediaType = MEDIA_TYPE_PLAIN;
         }
 
     }
 
     @Override
-    protected RequestBody buildRequestBody()
-    {
-        return RequestBody.create(mediaType, content);
+    protected RequestBody buildRequestBody() {
+        return RequestBody.create(content, mediaType);
     }
 
     @Override
-    protected Request buildRequest( RequestBody requestBody)
-    {
+    protected Request buildRequest(RequestBody requestBody) {
         return builder.post(requestBody).build();
     }
 
